pyproject.toml
[project]
name = "Social-Foraging-in-Dynamic-Environments-Balancing-Information-Channels"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    # RL & ML
    "vmas>=1.4.3",
    "tensordict>=0.6.2",
    "torchrl>=0.6.0",
    "scikit-learn",
    "opencv-python>=4.10.0.84",
    # Tools
    "ipykernel>=6.29.5",
    "matplotlib>=3.9",
    "mpltern",
    "seaborn>=0.13.2",
    "notebook>=7.4.2",
    "pillow>=11.1.0",
    "hydra-core>=1.3.2",
    "wandb[media]>=0.19.2",
    "tqdm>=4.67.1",
    "deap>=1.4.3",
    "pandas>=2.3.3",
    "streamlit",
    "imageio",
    "pytest>=7.4.0",
    "optuna>=4.7.0",
    "optuna-dashboard>=0.20.0",
]

[project.optional-dependencies]
cpu = [
    "torch>=2.5.1",
]
cu121 = [
    "torch>=2.5.1",
]

[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "cu121" },
    ],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu121", extra = "cu121" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
